enterprise_reports/requirements.txt
# Django and REST Framework
Django>=5.2.0
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
django-extensions>=4.1.0
django-filter>=25.1.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
 
# Auth
djangorestframework-simplejwt>=5.3.1

# Background Tasks
celery>=5.5.0
redis>=6.4.0

# Report Generation
reportlab>=4.4.0
openpyxl>=3.1.0
pandas>=2.3.0

# Email and Storage
sendgrid>=6.12.0
boto3>=1.40.0

# Utilities
python-decouple>=3.8.0
Pillow>=9.0.0

# Development
django-debug-toolbar>=4.0.0
